Avatars: prevent requesting the same avatar multiple times
by waiting that the 1st one finish and retriving from memory
by managing a list of retrieve in progress.

Because when opening a new repository with all the last commits belonging to
a very few  authors (extreme case only one), the revision grid, to display the avatars
request the same one for each line.
So it could end up with 10 times the same slow http avatar request on GitHub or Gravatar.